requirements.txt
qrcode[pil]
customtkinter
Pillow
pyperclip
PyPix
